+++ programs/encore/src/instructions/mod.rs
@@ -1,7 +1,10 @@
 pub mod event_create;
 pub mod event_update;
 pub mod ticket_mint;
+pub mod ticket_transfer;
 
 pub use event_create::*;
 pub use event_update::*;
 pub use ticket_mint::*;
+pub use ticket_transfer::*;
+